Split CivilWarService into CivilWarPointService, CivilWarPvpService and CivilWarRankService #29
