+++ src/main/java/com/projectswg/holocore/services/gameplay/gcw/GalacticCivilWarManager.java
@@ -1,11 +1,12 @@
 package com.projectswg.holocore.services.gameplay.gcw;
 
-import com.projectswg.holocore.services.gameplay.gcw.faction.FactionManager;
 import me.joshlarson.jlcommon.control.Manager;
 import me.joshlarson.jlcommon.control.ManagerStructure;
 
 @ManagerStructure(children = {
-		FactionManager.class
+		CivilWarPointService.class,
+		CivilWarPvpService.class,
+		CivilWarRankService.class,
 })
 public class GalacticCivilWarManager extends Manager {
