+++ src/test/java/com/projectswg/holocore/services/gameplay/gcw/TestCivilWarPvpService.java
@@ -25,7 +25,7 @@
  * along with Holocore.  If not, see <http://www.gnu.org/licenses/>.               *
  ***********************************************************************************/
 
-package com.projectswg.holocore.services.gameplay.gcw.faction;
+package com.projectswg.holocore.services.gameplay.gcw;
 
 import com.projectswg.common.data.encodables.tangible.PvpFaction;
 import com.projectswg.holocore.resources.support.objects.swg.creature.CreatureDifficulty;
@@ -33,15 +33,12 @@ import com.projectswg.holocore.test.runners.TestRunnerNoIntents;
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.time.LocalDate;
-import java.time.OffsetDateTime;
-
-public class TestCivilWarService extends TestRunnerNoIntents {
+public class TestCivilWarPvpService extends TestRunnerNoIntents {
 	
-	private final CivilWarService service;
+	private final CivilWarPvpService service;
 	
-	public TestCivilWarService() {
-		service = new CivilWarService();
+	public TestCivilWarPvpService() {
+		service = new CivilWarPvpService();
 	}
 	
 	@Test
@@ -71,47 +68,6 @@ public class TestCivilWarService extends TestRunnerNoIntents {
 	@Test
 	public void testPointsGranted() {
 		Assert.assertEquals(200, service.pointsGranted(10, (byte) 20));
-	}
-	
-	private int epochTime(LocalDate date) {
-		return (int) date.atStartOfDay(OffsetDateTime.now().getOffset()).toEpochSecond();
-	}
-	
-	@Test
-	public void testNextUpdateTime() {
-		LocalDate now = LocalDate.of(2018, 2, 19);	// It's a Monday
-		LocalDate rankDay = LocalDate.of(2018, 2, 23);	// It's a Friday, the exact time we rank up
-		LocalDate dayAfter = LocalDate.of(2018, 2, 24);	// It's a Saturday, 24 hours after rank up
-		LocalDate nextRankDay = LocalDate.of(2018, 3, 2);	// It's a Friday, exactly one week after first rank up
-		
-		int nowRankTime = epochTime(rankDay);
-		int nextRankTime = epochTime(nextRankDay);
-		
-		Assert.assertEquals(nowRankTime, service.nextUpdateTime(now));
-		Assert.assertEquals(nextRankTime, service.nextUpdateTime(rankDay));	// When we hit the scheduled rank time, the next update should be in a week
-		Assert.assertEquals(nextRankTime, service.nextUpdateTime(dayAfter));	// Next time should be in six days (from Saturday to Friday)
-	}
-	
-	@Test
-	public void testIsDecayRank() {
-		Assert.assertFalse(service.isDecayRank(6));
-		Assert.assertTrue(service.isDecayRank(7));
-	}
-	
-	@Test
-	public void testRankProgress() {
-		Assert.assertEquals(2.86f, service.rankProgress(10.0f, 20.0f, 7, 9000), 1);
-	}
-	
-	@Test
-	public void testIsRankDown() {
-		Assert.assertFalse(service.isRankDown(40, 50));
-		Assert.assertTrue(service.isRankDown(3, -10));
-	}
-	
-	@Test
-	public void testLeftoverPoints() {
-		Assert.assertEquals(6000, service.leftoverPoints(130, 20000));
 	}
 	
 }
